### 2017-08-11 150b7c8
Added DOI's to entries in docs/joss/paper.bib.
