+++ docs/joss/paper.bib
@@ -22,7 +22,8 @@
   number={2},
   pages={22--30},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MCSE.2011.37}
 }
 
 @article{van2014scikit,
@@ -32,7 +33,8 @@
   volume={2},
   pages={e453},
   year={2014},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi={10.7717/peerj.453}
 }
 
 @article{dalcin2011parallel,
@@ -43,7 +45,8 @@
   number={9},
   pages={1124--1139},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.advwatres.2011.04.013}
 }
 
 @article{kloeckner_pycuda_2012,
@@ -75,5 +78,6 @@
     year = {2013},
     publisher = {O'Reilly},
     title = {Python and HDF5},
-    author = {Andrew Collette}
+    author = {Andrew Collette},
+    isbn={9781491945018}
 }
