### 2017-08-11 702dd01
Bump version.txt to 0.6.0dev2
