+++ array_split/version.txt
@@ -1,1 +1,1 @@
-0.6.xdev0
+0.6.xdev2
