### 2017-08-12 #345 @Shane-J-Latham
Hi @arfon, thanks for taking the time to look at this.

> Could you confirm here that there is a research application

I would say it is a tool for "enabling research", particularly in HPC where
you want to calculate domain-decomposition for data-parallelism
(for large-image processing, say)...
If you don't think this is a fit for JOSS, happy to withdraw.

Thanks again.
