### 2017-08-15 #345 @arfon
Thanks @Shane-J-Latham. It sounds like this submission falls into the ['utility package' category](http://joss.theoj.org/about#submission_requirements) which _is_ acceptable as long as it's not a very small contribution.

👍 I'm happy to proceed with this submission.
