### 2017-08-16 #345 @dvalters
@arfon I think this is quite an interesting submission (I can see applications of it for research as a utility package). I'd be interested in reviewing it but would not be able to start until end of next week.
