### 2017-08-19 #345 @arfon
> I'd be interested in reviewing it but would not be able to start until end of next week.

OK no problem.
