### 2017-08-19 #373 @whedon
[REVIEW]: array_split: Multi-dimensional array partitioning
**Submitting author:** @Shane-J-Latham (<a href="http://orcid.org/0000-0001-8033-4665">Shane J. Latham</a>)
**Repository:** <a href="https://github.com/array-split/array_split" target ="_blank">https://github.com/array-split/array_split</a>
**Version:** 0.5.0
