### 2017-09-07 #373 @dvalters
# Review summary
@arfon @Shane-J-Latham

My review is quite short as I could find little to fault in the submission. The standard of the **array_split** software package is high. It is well documented and with a good level of testing. It provides a useful and substantial extension to the functionality of the basic NumPy array splitting functions. There are good applications for scientific parallel codes as the author suggests in the paper. The quality of the Python code is also high.

I have very few requests before the paper can be accepted, just minor issues with some of the references:

## References
Could you check that DOIs are provided for papers that have them. The Dalcin (2011) paper has a DOI available on the [Science Direct website](http://www.sciencedirect.com/science/article/pii/S0309170811000777), but it is not included in the `.bib` file. Similarly for the reference to scikit-image (https://www.ncbi.nlm.nih.gov/pubmed/25024921), and the NumPy Array reference (http://ieeexplore.ieee.org/document/5725236/).

Once these are fixed, I can recommend it be accepted into JOSS.

## Suggestions (not required for acceptance)
The documentation is a good standard, but it would be really useful if at somepoint diagrams could be added to illustrate how the arrays are being split with the different functions (I find I often have to sketch out what I want the arrays/halos to look like when working on problems like this). The writing as it stands is clear, however, and this is only a suggestion for future work and not required for acceptance in my opinion.
