### 2017-09-07 #373 @Shane-J-Latham
Thanks for the review @dvalters, much appreciated.

Yes, I neglected to include DOIs in the 0.5.0 version, have subsequently updated
them in the `0.5.x_rc` branch, see this [paper.bib](https://github.com/array-split/array_split/blob/0.5.x_rc/docs/joss/paper.bib). I'll create a 0.5.1 version,
which includes the corrected `paper.bib`, on acceptance for the DOI.

Some diagrams in the documentation would indeed add some clarity.
More generally, the ability to visualise the calculated *split* would also
be valuable. Could plot 1D, 2D and maybe 3D *splits* using [matplotlib](https://matplotlib.org/),
and for higher dimensions plot 2D slices at specified axis positions...
