### 2017-09-11 #373 @arfon
@Shane-J-Latham - please merge [this PR](https://github.com/array-split/array_split/pull/1) fixing up some issues with the bibtex. Also, then please could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
