### 2017-09-12 #373 @array-split
Thanks for the paper.bib syntax fixes @arfon. DOI is 10.5281/zenodo.889080
