### 2017-09-12 afcd142
Added DOI badge to README.rst.
